+++ src/ChunkPrinter.cpp
@@ -1,5 +1,6 @@
+#ifdef ARDUINO
 
-#include "ChunkPrinter.h"
+  #include "ChunkPrinter.h"
 
 ChunkPrinter::ChunkPrinter(PsychicResponse* response, uint8_t* buffer, size_t len) : _response(response),
                                                                                      _buffer(buffer),
@@ -18,8 +19,7 @@ size_t ChunkPrinter::write(uint8_t c)
   esp_err_t err;
 
   // if we're full, send a chunk
-  if (_pos == _length)
-  {
+  if (_pos == _length) {
     _pos = 0;
     err = _response->sendChunk(_buffer, _length);
 
@@ -36,16 +36,14 @@ size_t ChunkPrinter::write(const uint8_t* buffer, size_t size)
 {
   size_t written = 0;
 
-  while (written < size)
-  {
+  while (written < size) {
     size_t space = _length - _pos;
     size_t blockSize = std::min(space, size - written);
 
     memcpy(_buffer + _pos, buffer + written, blockSize);
     _pos += blockSize;
 
-    if (_pos == _length)
-    {
+    if (_pos == _length) {
       _pos = 0;
 
       if (_response->sendChunk(_buffer, _length) != ESP_OK)
@@ -58,8 +56,7 @@ size_t ChunkPrinter::write(const uint8_t* buffer, size_t size)
 
 void ChunkPrinter::flush()
 {
-  if (_pos)
-  {
+  if (_pos) {
     _response->sendChunk(_buffer, _pos);
     _pos = 0;
   }
@@ -69,11 +66,9 @@ size_t ChunkPrinter::copyFrom(Stream& stream)
 {
   size_t count = 0;
 
-  while (stream.available())
-  {
+  while (stream.available()) {
 
-    if (_pos == _length)
-    {
+    if (_pos == _length) {
       _response->sendChunk(_buffer, _length);
       _pos = 0;
     }
@@ -84,3 +79,5 @@ size_t ChunkPrinter::copyFrom(Stream& stream)
   }
   return count;
 }
+
+#endif // ARDUINO